+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.16.0
+* BREAKING CHANGE: `payload` is now `payloadPrefix`
+* BREAKING CHANGE: `mask` is now `payloadMask`
 * Bump bluez to 0.8.3
 * Do not normalize UUIDs on native side
 * Reverse _permissionStatus return values in example app